# platform/backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PureData API",
    "description": "AI训练数据生成平台 - 高性能数据生成服务\n\n## 核心功能\n- **多领域数据生成**：人工智能、医疗、金融、劳动合同、法律、教育、电商、科技\n- **质量门控系统**：高质量(≥0.85)、中等质量(0.70-0.85)、低质量(<0.70)\n- **智能去重**：MinHash近似去重\n- **数据血缘追溯**：完整的数据来源追踪\n- **异常检测**：自动检测和修复数据异常\n- **Calibrated Mixup增强**：提升低分样本质量\n\n## 认证方式\n- 用户认证：Bearer Token 或 Cookie\n- 管理员认证：独立的Token系统\n\n## 成本说明\n- 知识库数据：免费\n- API生成数据：约0.3元/百万token（qwen-turbo）",
    "version": "2.2.0",
    "contact": {
      "name": "PureData Team",
      "email": "support@puredata.ai"
    },
    "license": {
      "name": "Commercial License",
      "url": "https://puredata.ai/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发服务器"
    },
    {
      "url": "https://api.puredata.ai",
      "description": "生产服务器"
    }
  ],
...
  "paths": {
...
    "/generate": {
      "post": {
        "tags": ["数据生成"],
        "summary": "生成数据",
        "description": "生成指定领域和数量的数据",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["domain", "count"],
                "properties": {
                  "domain": {
                    "type": "string",
                    "enum": ["人工智能", "医疗", "金融", "劳动合同", "法律", "教育", "电商", "科技"],
                    "example": "人工智能"
                  },
                  "count": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100000,
                    "example": 100
                  },
                  "quality_level": {
                    "type": "string",
                    "enum": ["high_quality", "medium_quality", "low_quality"],
                    "default": "high_quality",
                    "description": "质量等级：high_quality(≥0.85), medium_quality(0.70-0.85), low_quality(<0.70)"
                  },
                  "format": {
                    "type": "string",
                    "enum": ["json", "jsonl", "csv"],
                    "default": "json"
                  },
                  "use_api": {
                    "type": "boolean",
                    "default": false,
                    "description": "是否使用API生成（付费）"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "生成任务已创建",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "task_id": {"type": "string", "example": "abc123"},
                    "message": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {"description": "未授权"},
          "400": {"description": "参数错误"}
        }
      }
    },
    "/task/{task_id}": {
      "get": {
        "tags": ["数据生成"],
        "summary": "获取任务状态",
        "description": "查询数据生成任务的状态和进度",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "task_id": {"type": "string"},
                    "status": {"type": "string", "enum": ["pending", "processing", "completed", "failed"]},
                    "progress": {"type": "integer", "minimum": 0, "maximum": 100},
                    "filename": {"type": "string"},
                    "count": {"type": "integer"},
                    "quality_stats": {
                      "type": "object",
                      "properties": {
                        "avg_score": {"type": "number"},
                        "high_quality_count": {"type": "integer"},
                        "medium_quality_count": {"type": "integer"},
                        "low_quality_count": {"type": "integer"}
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "用户JWT Token"
      },
      "adminAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "管理员JWT Token"
      }
    },
...
  }
}